refact: cebab-case & css modules
