+++ src/app.module.css
@@ -1,0 +1,9 @@
+.app {
+  min-width: 375px;
+  margin: auto;
+  display: flex;
+  flex-flow: column nowrap;
+  align-items: center;
+  min-height: 100vh;
+  box-sizing: border-box;
+}
