--- src/components/AppHeader/AppHeader.module.css
@@ -1,7 +1,0 @@
-.AppHeader {
-  width: 100%;
-  min-height: 88px;
-  background-color: var(--color-background-modal);
-  display: flex;
-  align-items: center;
-}
