--- src/components/BurgerComponent/BurgerComponent.css
@@ -1,6 +1,0 @@
-.BurgerComponent {
-  display: flex;
-  flex-flow: row nowrap;
-  gap: 12px;
-  align-items: center;
-}
