--- src/components/BurgerComponent/BurgerComponent.js
@@ -1,14 +1,0 @@
-import { ConstructorElement } from '@ya.praktikum/react-developer-burger-ui-components';
-import { getIcons } from '../../utils/utils';
-import './BurgerComponent.css';
-
-export const BurgerComponent = (props) => {
-  const { withDrag, ...otherPorps } = props;
-
-  return (
-    <div className="BurgerComponent">
-      {withDrag && getIcons('primary')['drag']}
-      <ConstructorElement {...otherPorps} />
-    </div>
-  );
-};
