--- src/components/BurgerComponents/BurgerComponents.css
@@ -1,6 +1,0 @@
-.BurgerComponents {
-  display: flex;
-  flex-flow: column;
-  align-items: end;
-  gap: 16px;
-}
