--- src/components/BurgerConstructor/BurgerConstructor.css
@@ -1,3 +1,0 @@
-.BurgerConstructor {
-  justify-self: end;
-}
