--- src/components/BurgerConstructorTotal/BurgerConstructorTotal.css
@@ -1,5 +1,0 @@
-.BurgerConstructorTotal {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-}
