--- src/components/BurgerIngredients/BurgerIngredients.module.css
@@ -1,3 +1,0 @@
-.BurgerIngridients {
-  width: 100%;
-}
