--- src/components/Nav/Nav.module.css
@@ -1,9 +1,0 @@
-.Nav {
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-
-  display: flex;
-  flex-flow: row;
-  align-items: center;
-}
