--- src/components/NavItem/NavItem.module.css
@@ -1,21 +1,0 @@
-.NavItem {
-  display: flex;
-  flex-flow: row nowrap;
-  gap: 8px;
-}
-
-.NavItem:first-of-type {
-  padding-left: 0;
-}
-
-.NavItem:last-of-type {
-  padding-right: 0;
-}
-
-.NavItem_align_right {
-  margin-left: auto;
-}
-
-.NavItem_align_left {
-  margin-right: auto;
-}
