+++ src/components/app-header/app-header.js
@@ -1,0 +1,12 @@
+import { Nav } from '../nav';
+import s from './app-header.module.css';
+
+export const AppHeader = () => {
+  return (
+    <header className={s.appheader}>
+      <div className="page__section">
+        <Nav />
+      </div>
+    </header>
+  );
+};
