+++ src/components/app-header/app-header.module.css
@@ -1,0 +1,7 @@
+.appheader {
+  width: 100%;
+  min-height: 88px;
+  background-color: var(--color-background-modal);
+  display: flex;
+  align-items: center;
+}
