+++ src/components/burger-component/burger-component.module.css
@@ -1,0 +1,6 @@
+.burgerComponent {
+  display: flex;
+  flex-flow: row nowrap;
+  gap: 12px;
+  align-items: center;
+}
