+++ src/components/burger-component/index.js
@@ -1,0 +1,3 @@
+import { BurgerComponent } from './burger-component';
+
+export { BurgerComponent };
