+++ src/components/burger-components/burger-components.module.css
@@ -1,0 +1,6 @@
+.burgerComponents {
+  display: flex;
+  flex-flow: column;
+  align-items: end;
+  gap: 16px;
+}
