+++ src/components/burger-components/index.js
@@ -1,0 +1,3 @@
+import { BurgerComponents } from './burger-components';
+
+export { BurgerComponents };
