+++ src/components/burger-constructor-total/burger-constructor-total.module.css
@@ -1,0 +1,5 @@
+.burgerConstructorTotal {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+}
