+++ src/components/burger-constructor-total/index.js
@@ -1,0 +1,3 @@
+import { BurgerConstructorTotal } from './burger-constructor-total';
+
+export { BurgerConstructorTotal };
