+++ src/components/burger-constructor/burger-constructor.module.css
@@ -1,0 +1,3 @@
+.burgerConstructor {
+  justify-self: end;
+}
