+++ src/components/burger-constructor/index.js
@@ -1,0 +1,3 @@
+import { BurgerConstructor } from './burger-constructor';
+
+export { BurgerConstructor };
