+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -1,0 +1,3 @@
+.burgerIngridients {
+  width: 100%;
+}
