+++ src/components/burger-ingredients/index.js
@@ -1,0 +1,2 @@
+import { BurgerIngredients } from './burger-ingredients';
+export { BurgerIngredients };
