+++ src/components/main/index.js
@@ -1,0 +1,3 @@
+import { Main } from './main';
+
+export { Main };
