+++ src/components/main/main.module.css
@@ -1,0 +1,10 @@
+.main {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  justify-items: center;
+  gap: 40px;
+  padding: 0 40px;
+  box-sizing: border-box;
+  max-width: calc(1240px + var(--gutter) * 2);
+  width: 100%;
+}
