+++ src/components/nav-item/index.js
@@ -1,0 +1,3 @@
+import { NavItem } from './nav-item';
+
+export { NavItem };
