+++ src/components/nav/index.js
@@ -1,0 +1,3 @@
+import { Nav } from './nav';
+
+export { Nav };
