+++ src/components/nav/nav.module.css
@@ -1,0 +1,9 @@
+.nav {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+
+  display: flex;
+  flex-flow: row;
+  align-items: center;
+}
