+++ src/utils/index.js
@@ -1,0 +1,2 @@
+import { getIcons } from './utils';
+export { getIcons };
